--- src/components/organisms/SidebarList/index.ts
@@ -1,1 +1,0 @@
-export { SidebarList } from './SidebarList';
